QHttp2Connection: Avoid warning about different signedness
(...)/qtbase/src/testlib/qtestcase.h:626:34: warning: comparison of integers of different signs: 'const long long' and 'const unsigned long' [-Wsign-compare]
        return compare_helper(t1 == t2, "Compared values are not the same",
                              ~~ ^  ~~
(...)/home/qtrob/dev/src/qt-dev_08.14-base_imageformats_svg/qtbase/tests/auto/network/access/qhttp2connection/tst_qhttp2connection.cpp:490:9: note: in instantiation of function template specialization 'QTest::qCompare<long long, unsigned long>' requested here
        QCOMPARE(writtenN, buffer.size());
        ^
(...)/home/qtrob/dev/src/qt-dev_08.14-base_imageformats_svg/qtbase/src/testlib/qtestcase.h:83:17: note: expanded from macro 'QCOMPARE'
    if (!QTest::qCompare(actual, expected, #actual, #expected, __FILE__, __LINE__))\
                ^

+++ tests/auto/network/access/qhttp2connection/tst_qhttp2connection.cpp
@@ -487,7 +487,7 @@ void tst_QHttp2Connection::testBadFrameSize()
         buffer[2] = size;
 
         auto writtenN = connection->getSocket()->write(reinterpret_cast<const char *>(&buffer[0]), buffer.size());
-        QCOMPARE(writtenN, buffer.size());
+        QCOMPARE(writtenN, qint64(buffer.size()));
     }
 
     QCOMPARE(rstClientSpy.wait(), rst_received);
